+++ android/src/main/kotlin/dev/yanshouwang/camerax/CameraXPlugin.kt
@@ -5,12 +5,17 @@ import androidx.annotation.NonNull
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.embedding.engine.plugins.activity.ActivityAware
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+import io.flutter.plugin.common.EventChannel
+import io.flutter.plugin.common.MethodChannel
+import io.flutter.view.TextureRegistry
 
 /** CameraXPlugin */
 class CameraXPlugin : FlutterPlugin, ActivityAware {
     private var flutter: FlutterPlugin.FlutterPluginBinding? = null
     private var activity: ActivityPluginBinding? = null
     private var handler: CameraXHandler? = null
+    private var method: MethodChannel? = null
+    private var event: EventChannel? = null
 
     override fun onAttachedToEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
         this.flutter = binding
@@ -21,9 +26,12 @@ class CameraXPlugin : FlutterPlugin, ActivityAware {
     }
 
     override fun onAttachedToActivity(binding: ActivityPluginBinding) {
-        handler = CameraXHandler(binding.activity, this.flutter!!.binaryMessenger, this.flutter!!.textureRegistry)
-        handler!!.startListening()
         activity = binding
+        handler = CameraXHandler(activity!!.activity, flutter!!.textureRegistry)
+        method = MethodChannel(flutter!!.binaryMessenger, "yanshouwang.dev/camerax/method")
+        event = EventChannel(flutter!!.binaryMessenger, "yanshouwang.dev/camerax/event")
+        method!!.setMethodCallHandler(handler)
+        event!!.setStreamHandler(handler)
         activity!!.addRequestPermissionsResultListener(handler!!)
     }
 
@@ -33,9 +41,12 @@ class CameraXPlugin : FlutterPlugin, ActivityAware {
 
     override fun onDetachedFromActivity() {
         activity!!.removeRequestPermissionsResultListener(handler!!)
-        activity = null
-        handler!!.stopListening()
+        event!!.setStreamHandler(null)
+        method!!.setMethodCallHandler(null)
+        event = null
+        method = null
         handler = null
+        activity = null
     }
 
     override fun onDetachedFromActivityForConfigChanges() {