+++ ios/Classes/CameraXPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface CameraXPlugin : NSObject<FlutterPlugin>
+@end