+++ lib/mlkit.dart
@@ -1,0 +1,3 @@
+library mlkit;
+
+export 'src/mlkit/barcode.dart';