+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.2
+
+- Add CameraWidget and CameraController on iOS.
+- Add barcode stream (detected by MLKit) instead of image stream.
+- Update camerax example with barcode detection.
+
 ## 0.0.1
 
 - Add CameraWidget and CameraController on Android.